refactor get_replace_regex_dict_state
